requirements.txt
matplotlib
numpy
pandas
odfpy
pytz
requests
tabulate